Read entire timestep's value from HDF5 files
When reading a quantity `vals` from an HDF5 file with `quantity_values`,
indexing on `vals` will read the value from the file. Before, the
resulting `vals[i]` would still be a "view" into the file. If `vals[i]`
was a large array in the previous version, iterating over the array
would be extremely slow due to the IO bottleneck.