Allow splitting curves and partitioning later
Similar to `separate(segments, cuts)`, a method has been added to
separate a curve at certain cuts. When the curve is `partition`ed, the
result can be passed to `separate` to retrieve a tuple of `Segments`.

This allows the user to cut the curve in one coordinate system,
transform the curve, then finally partition the curve.