Stop `quantity_values` from referencing the HDF5 file
Before, `quantity_values` would reference the HDF5 file such that the
entire value wasn't read into memory. This would definitely be a
source of confusion, so by default it will now read the entire value
into memory.

In the future, another function or method could be added to provide
the old functionality.